Implemented Devblog and news commands, added comments
Implemented an xml scrape of FP's rss feed to fetch the most recent devblog. Added html scrape of Rustafied's homepage to get a title & description of their current news. In the future, as soon as either of the sites update I would like CamBot to make a discord post. Additionally, I started adding comments to improve future readability. Not shown here, but I also am creating the infrastructure for the item database. I believe all tables and relations are in order I just have to add the hundreds of items to the database.