Modularized BS4 scraping, moved back to Rustlabs data
After trying to get an SQL server set up with that many items, I decided to reevaluate my approach to getting the item info. As Rustlabs already had all of the info and I was using it to populate my database anyway, I decided to attempt scraping their site again. I was able to take a much more sophisticated approach to getting the data I needed. The SQL server was way to cumbersome when it came to entering data with multiple relations. If I were to set up my own item database, it would be using something link MondoDB or something with less restrictions and structure. The craftcalc now works and all scraping functionality now calls a modularized function