Added background task to check for site updates
Added a background task loop that will check the Rust item store, rustafied's website, and the facepunch devblog every 15 minutes for an update. If there is any update, CamBot will post a corresponding announcement to all servers it is in. Additionally, if there are new skins added to the item store, their data will be added to the SQL server