Greatly improved skin database. Added skin typing
Using a separate program, I updated the skin database to hold the types of skins. Additionally, the initial prices of over half of the skins in the database have been make 100% accurate. The data containing these initial prices, especially early ones, is hard to come by and was found in 3 separate ways which I had to programatically incorporate into a script that updated the database. From youtube videos, to my own steam purchase history, to using the wayback machine, updating the prices was definitely a hassle.