Bug fixes for updating new skins from store
I changed inserting skins into the SQL database to be more object-oriented. To do this, I created a skin class used to store data for each skin to be inserted. This led to a bug where I tried to insert the skin object instead of its name into the "skins.txt" text file. Additionally, with the addition of skin types to the SQL database, I forgot to handle the exception where an item has no type. Both of these errors should be fixed now