Optimized site checking
Instead of getting all items from the rust store page and only comparing the first one, only the first entry is retrieved from the site and compared to check for changes. This is much more efficient because the code I was previously using was updated to predict each price when retrieved from the store page which greatly increased the time cost of calling that function.