Added early iteration of modular images
Implemented an early method of how I want to display skin data for the !rustitems command. This method uses PIL to create a modular image for each skin with its respective data. However, discord downscales images that are too large so I have had to create an image for every item. This system will need to be researched further to perfect it but it is in a good enough place to include. Additionally, I have added a cache system for the !rustitems command. Because I am now including predicted prices, which are expensive to generate, a cache makes it so that these predictions only need to be generated once.