Improved output
Improved output for rustitems, experiment, droptable, lootfrom, stats, and status commands. Implemented a function to split a list of strings into columns that was used to improve output for experiment, droptable, and lootfrom cmds. Removed unnecessary if statement when splitting long strings. Added general info to stats command along with an image of the item.