Integrated dynamic embeds and custom emojis
Added two features that were being tested in a separate environment. Cambot will now add custom emojis to a discord server, given enough emoji slots are available. These emojis will be used for anything that features crafting/materials such as the !craftcalc command.

Additionally, I have moved away from using custom images with PIL for displaying the current rust items as they were still ugly UI-wise and didn't come out the way I initially wanted. This is because I failed to account for discord downscaling images to a height of ~300px no matter the original size. To circumvent this I initially uploaded an image for each item but that wasn't very appealing and it looked bad on mobile. To fix this, I implemented dynamic embeds which will display 1 item at a time but by reacting to the message, a user can scroll through all skins. I implemented an SQL backend such that older messages that are stored in the server will still be navigable as well