Migrated to SQLite, minor UI tweaks & bug fixes
Added images to lootfrom command. Changed layout of default output for smelting and droptable commands

Fixed bugs with craftcalc and repair when items can't be crafted/repaired

Migrated data from an AWS RDB server to SQLite as I didn't want to pay money and I don't need the concurrent access as the only one who should have access to the db is the bot.

Added a small block of code at the end of command execution that calculates the time elapsed and stores it in the SQLite file. From there, I have a separate python script that aggregates the execution times and displays an average for each command. With this I intend to measure optimization efficacy using current times as a baseline. This will prove useful when migration from web scraping to full database dependency begins