Optimized most commands, added commands
Finished optimization of all Cambot commands. Some commands are dynamic in nature and thus will always need to scrape a webpage. These commands were not optimized and have the same execution time.

With the migration to an SQLite database, I created a script to update the entire database in one pass. This script gets all items/buildings on rustlabs and updates all tables with information on each item. If an item does not have information for a specific table, it is skipped. When creating the database I decided not to normalize it as this shift to a database in the first place was entirely for performance and thus having multiple table joins per query was not optimal. The script itself runs in ~20minutes and should only need to be executed every month at max. I also added a command to update the database which will send out an automated tweet alerting users of the downtime as command are not available while the update script is running.

After migrating to the database and optimizing my code I was able to bring execution times down to as low as 10% of what they previously were. On average, commands were measured to execute in 25% of the time it took previously. 

Additionally, I overhauled the dynamic embed system to allow for more message types to be stored in the backend. Now any message type should be able to be dynamic.

Added the durability command. This command displays the durability of any building/deployable in Rust. It will show how many explosives/tools are required to break the item.

Added the mix command for Rust's new mixing table. This command displays recipes for all items craftable in the mixing table. This is separate from craftcalc as each slot in the mixing table is static and the same item can be in different slots.

Added the composting command that displays the compost yields for all applicable items. You can also look up the compost amount for a single item.

Changed the gen_emoji command to return the input string if no emoji is found, this is more efficient.

Added a format_embed function that takes all information about an embed and formats the fields depending on the number of them.